initial version of the project
We want to use the teensy41 as an intermediate device that is both
useful for acting as a secondary onboard cache but also a high
performance interface for different kinds of peripherals as well.

The onboard 16 megabytes of PSRAM are used as a 16M of ram that is
accessed via caches.