Seed the cache on startup to eliminate the valid bit
On startup, the cache is in an invalid state. We load the first 4k of
memory into the cache to make sure we always have some sort of valid
state to work with. Since I will never have to deal with an MMU, the
idea of an invalid state just won't exist.

This has the side effect of simplifying the checking process by a little
bit