Introduce a new skewed cache index computation layout
We eliminate the right shifts by turning the lower byte into the base
index and the upper 4 bits of the lowest byte into an offset. This
prevents collisions and poor usage of the direct mapped cache. It
reduces the index computation overhead by a massive factor. INtead of
requring 6 instructions, it now requires two instructions.